% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_weight_preferences.R
\name{biomass_weight_preferences}
\alias{biomass_weight_preferences}
\title{A function to calculate carbon and nitrogen fluxes in the food web.}
\usage{
biomass_weight_preferences(usin)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}
}
\value{
The community that you are analyzing with the biomass weighted preferences includes in the interaction matrix instead of the raw unweighted preferences.
}
\description{
A function to calculate carbon and nitrogen fluxes in the food web.
}
\examples{
biomass_weight_preferences(intro_comm)
}
