% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_foodweb.R
\name{build_foodweb}
\alias{build_foodweb}
\title{A function to compile the food web from simple data inputs}
\usage{
build_foodweb(feeding, properties)
}
\arguments{
\item{feeding}{A data frame listing the feeding relationships in the food web. Only needs to contain the nodes that have prey items along with their names. Three columns in this data frame must be: Predator, Prey, and Preference. The Predator column is the name of the predator, the Prey column is the name of the prey, and the Preference is the preference that the predator has for that prey item after correcting for abundance. Preference values are relative, so for each predator the preference value can be any number and all that matters is its proportion of the total of all preference values given to that predator. Preference is meaningless if a predator only has one prey item. A good default value is 1 for everything if you don't want to set preferences beyond prey abundance.}

\item{properties}{A data frame listing the properties or parameters in the food web for each element. Must contain the following columns: ID, Element,Parameter, Value. For Carbon, the following parameters are required for each node: a, E, Q, canIMM, d, B, Detritusrecycling, isDetritus, and isPlant. For every other element the following parameters are required for each node: a, Q, canIMM. You can also include the parameters p and Emin for each element and Ehat for carbon. If these are not present, the model sets them to be trivial values that do not affect the calculations.}
}
\value{
A community that is compatible with the functions of the package. This is a list with the feeding matrix imat first and the properties prop second. The properties list contains a data frame for every element.
}
\description{
A function to compile the food web from simple data inputs
}
