% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_foodweb_inputs.R
\name{build_foodweb_inputs}
\alias{build_foodweb_inputs}
\title{A function to create data tables for food web inputs}
\usage{
build_foodweb_inputs(nodenames, elements)
}
\arguments{
\item{nodenames}{A vector of node names.}

\item{elements}{A vector of elements used in the model.}
}
\value{
A list of data tables for feeding matrices.
}
\description{
A function to create data tables for food web inputs
}
