% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Cijfcn}
\alias{Cijfcn}
\title{A utility function to calculate the consumption rate of each species on all prey assuming a type I functional response.}
\usage{
Cijfcn(usin, h = NA)
}
\arguments{
\item{usin}{The community on which feeding rate calculations are made.}

\item{h}{The matrix of the handling times that you want to use a type II functional response. If NA, you are using a Type I functional response. If you want some Type I and some Type II functional responses, include the matrix and put values of 0 for all Type I functional responses.}
}
\value{
A matrix of consumption rates with units set by the the biomass input units in biomass and time.
}
\description{
A utility function to calculate the consumption rate of each species on all prey assuming a type I functional response.
}
\examples{
Cijfcn(intro_comm)
}
