% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comana.R
\name{comana}
\alias{comana}
\title{A function to calculate carbon and nitrogen fluxes in the food web.}
\usage{
comana(usin, shuffleTL = FALSE, rmzeros = TRUE, eqmtolerance = NA)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}

\item{shuffleTL}{A Boolean stating whether the community should be sorted.}

\item{rmzeros}{A Boolean determining whether trophic species with zero biomass should be removed from the community before analysis.}

\item{eqmtolerance}{A value used to set the equilibrium tolerance for the food web verification. If NA, the default value used by the function all.equal is used, which is approximately 1.5e-8.}
}
\value{
A list of consumption rates, carbon mineralization, nitrogen mineralization, carbon and nitrogen consumption rates, and the modified community if zeros where removed or sorting occurred.
}
\description{
A function to calculate carbon and nitrogen fluxes in the food web.
}
\examples{
comana(intro_comm)
}
