% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_assimilation.R
\name{correct_assimilation}
\alias{correct_assimilation}
\title{A function to reduce growth by increasing respiration.}
\usage{
correct_assimilation(
  usin,
  output_type = TRUE,
  biomass_weight_preference = FALSE
)
}
\arguments{
\item{usin}{The input community.}

\item{output_type}{Should the nutrient limitation be printed (TRUE) or included in the output as a second object (FALSE)?}

\item{biomass_weight_preference}{Should the preference matrix be weighted by biomass inside this function? Default, False, assumes that you have already done this with the function biomass_weight_preferences or don't want to weight by biomass.}
}
\value{
The modified community with a lower carbon assimilation efficiency when necessary.
}
\description{
A function to reduce growth by increasing respiration.
}
