% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_respiration.R
\name{correct_respiration}
\alias{correct_respiration}
\title{A function to reduce growth by increasing respiration.}
\usage{
correct_respiration(usin, output_type = TRUE)
}
\arguments{
\item{usin}{The input community.}

\item{output_type}{Should the nutrient limitation be printed (TRUE) or included in the output as a second object (FALSE)?}
}
\value{
The modified community with a higher respiration rate.
}
\description{
A function to reduce growth by increasing respiration.
}
