% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_to_p.R
\name{E_to_p}
\alias{E_to_p}
\title{A function to convert between respiration by biomass carbon (E) and consumption (p).}
\usage{
E_to_p(usin)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}
}
\value{
The community where E is now zero and p is not 1.
}
\description{
A function to convert between respiration by biomass carbon (E) and consumption (p).
}
\examples{
E_to_p(intro_comm)
}
