% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{feedinglist}
\alias{feedinglist}
\title{The feeding list for an example community}
\format{
A single dataframe with columns Predator, Prey, Preference, aCarbon, aNitrogen, aPhosphorus, and aCalcium. The columns beginning in a are the assimilation efficiencies for each trophic interaction for that element.
\describe{
  \item{feedinglist}{The entire dataframe.}
}
}
\source{
Example not based on real empirical data.
}
\usage{
feedinglist
}
\description{
The community contains 5 nodes and 4 chemical elements.
}
\keyword{datasets}
