% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebode.R
\name{foodwebode}
\alias{foodwebode}
\title{A function to simulation the food webs away from equilibrium.}
\usage{
foodwebode(t, y, pars)
}
\arguments{
\item{t}{The ODE time.}

\item{y}{The ODE simulation start.}

\item{pars}{The ODE parameters.}
}
\value{
The changes in each node biomass along with parameters and mineralization rates.
}
\description{
A function to simulation the food webs away from equilibrium.
}
\details{
The food web model for simulating over time: requires \code{y} inputs and \code{pars} parameters from the \code{\link{getPARAMS}} function.
}
