% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPARAMS.R
\name{getPARAMS}
\alias{getPARAMS}
\title{A function to get the parameters for a food web model.}
\usage{
getPARAMS(
  usin,
  DIETLIMITS = NA,
  diet_correct = TRUE,
  Conly = FALSE,
  densitydependence = NA,
  functionalresponse = NA,
  returnnet = FALSE
)
}
\arguments{
\item{usin}{The community you want to simulate.}

\item{DIETLIMITS}{The diet limits matrix for the stoichiometry correction (proportion of diet)?}

\item{diet_correct}{Boolean: Does the organism correct it's diet?}

\item{Conly}{Boolean: Is the model meant for carbon only?}

\item{densitydependence}{Which nodes have density dependence? NA default means none of them do. Should be a vector of 0 (no DD) and 1 (DD) for each node.}

\item{functionalresponse}{The type of functional response to be used in the model simulation. Either NA to signify a Type I functional response or a matrix of values of the handling time to use a Type II functional response.}

\item{returnnet}{A Boolean to determine if the goal is to return the net change in the food web. Used to check equilibrium.}
}
\value{
A list with two elements: (1) a vector of parameters to run the model away from equilibrium and (2) a vector of equilibrium biomasses that can be modified and passed to the simulator.
}
\description{
A function to get the parameters for a food web model.
}
\details{
A function to get the parameters of a food web model for simulation purposes. It does not correct stoichiometry, so the user must do this beforehand if they want.
}
\examples{
# Basic call.
getPARAMS(intro_comm)
}
