% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{intro_comm2}
\alias{intro_comm2}
\title{The very simple 3-species example community produced.}
\format{
A list containing the feeding matrix and properties
\describe{
  \item{imat}{The feeding matrix.}
  \item{prop}{A list of property databases for each element. There are two parts to this list. The general part includes parameters at the node level and the assimilation part refers to the feeding-specific assimilation efficiency.}
}
}
\source{
Example not based on real empirical data.
}
\usage{
intro_comm2
}
\description{
The community contains 3 nodes and 2 chemical elements.
}
\keyword{datasets}
