% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_flux_view.R
\name{node_flux_view}
\alias{node_flux_view}
\title{A function to convert between respiration by biomass carbon (E) and consumption (p).}
\usage{
node_flux_view(usin, node = NULL, Element = NULL, mfrow_plot = c(1, 1))
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}

\item{node}{The node in the community whose fluxes you want to view. NULL is the default and means all of them. You can also include a vector of the ones that you want.}

\item{Element}{The element whose fluxes you want to view. NULL is the default and means all of them. You can also include a vector of the ones that you want.}

\item{mfrow_plot}{Two number vector saying the number of rows and columns for each plot page. Defaults to c(1,1).}
}
\value{
The community where E is now zero and p is not 1.
}
\description{
A function to convert between respiration by biomass carbon (E) and consumption (p).
}
\examples{
node_flux_view(intro_comm, node = "Pred", Element = "Carbon")
}
