% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_to_E.R
\name{p_to_E}
\alias{p_to_E}
\title{A function to convert between respiration by consumption (p) and biomass carbon (E).}
\usage{
p_to_E(usin)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}
}
\value{
The community where p is now 1 and E is not 0.
}
\description{
A function to convert between respiration by consumption (p) and biomass carbon (E).
}
\examples{
p_to_E(E_to_p(intro_comm))
}
