% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{properties_example1}
\alias{properties_example1}
\title{The properties database for an example community}
\format{
A single dataframe with columns ID, Element, Parameter, and Value.
\describe{
  \item{properties_example1}{The entire dataframe.}
}
}
\source{
Example not based on real empirical data.
}
\usage{
properties_example1
}
\description{
The community contains 5 nodes and 4 chemical elements.
}
\keyword{datasets}
