% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{removenodes}
\alias{removenodes}
\title{Remove nodes from community.}
\usage{
removenodes(usin, toremove)
}
\arguments{
\item{usin}{The community from which to remove nodes.}

\item{toremove}{A vector of nodes to remove using their names.}
}
\value{
The community without the removed nodes.
}
\description{
Remove nodes from community.
}
\examples{
removenodes(intro_comm, c("Pred"))
}
