% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_necro_feces.R
\name{sep_necro_feces}
\alias{sep_necro_feces}
\title{A function to calculate carbon and nitrogen fluxes in the food web.}
\usage{
sep_necro_feces(usin)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list. Make biomass_weighted feeding and stoichiometric correction first!}
}
\value{
The same community returned with the necromass and feces separated out from the detritus pools into which they original went.
}
\description{
A function to calculate carbon and nitrogen fluxes in the food web.
}
\examples{
sep_necro_feces(intro_comm)
}
