% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comana.R
\name{comana}
\alias{comana}
\title{A function to calculate carbon and nitrogen fluxes in the food web.}
\usage{
comana(usin, biomass_weight_preference = FALSE)
}
\arguments{
\item{usin}{The community that you are analyzing: contains a matrix of interactions and a data frame of properties in a list.}

\item{biomass_weight_preference}{Should the preference matrix be weighted by biomass inside this function? Default, False, assumes that you have already done this with the function biomass_weight_preferences or don't want to weight by biomass.}
}
\value{
A list of consumption rates, carbon mineralization, nitrogen mineralization, carbon and nitrogen consumption rates, and the modified community if zeros where removed or sorting occurred.
}
\description{
A function to calculate carbon and nitrogen fluxes in the food web.
}
\examples{
comana(intro_comm)
}
