% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPARAMS.R
\name{getPARAMS}
\alias{getPARAMS}
\title{A function to get the parameters for a food web model.}
\usage{
getPARAMS(
  usin,
  DIETLIMITS = NA,
  diet_correct = TRUE,
  Conly = FALSE,
  densitydependence = NA,
  functionalresponse = NA,
  externalinputs = NA,
  inorganicinputs = NA,
  inorganicloss = NA,
  returnnet = FALSE
)
}
\arguments{
\item{usin}{The community you want to simulate.}

\item{DIETLIMITS}{The diet limits matrix for the stoichiometry correction (proportion of diet)?}

\item{diet_correct}{Boolean: Does the organism correct it's diet?}

\item{Conly}{Boolean: Is the model meant for carbon only?}

\item{densitydependence}{Which nodes have density dependence? NA default means none of them do. Should be a vector of 0 (no DD) and 1 (DD) for each node.}

\item{functionalresponse}{The type of functional response to be used in the model simulation. Either NA to signify a Type I functional response or a matrix of values of the handling time to use a Type II functional response.}

\item{externalinputs}{A matrix of parameters with the rows being the nodes in the food web and the columns being the chemical elements. Non-zero inputs should match the stoichiometry of non-detritus pools. Detritus inputs should meet demands of the food web.}

\item{inorganicinputs}{A vector of inputs for the inorganic nutrients.}

\item{inorganicloss}{A vector of loss rates for the inorganic nutrients.}

\item{returnnet}{A Boolean to determine if the goal is to return the net change in the food web. Used to check equilibrium.}
}
\value{
A list with two elements: (1) a vector of parameters to run the model away from equilibrium and (2) a vector of equilibrium biomasses that can be modified and passed to the simulator.
}
\description{
A function to get the parameters for a food web model.
}
\details{
A function to get the parameters of a food web model for simulation purposes. It does not correct stoichiometry, so the user must do this beforehand if they want.
}
\examples{
# Basic call.
ei = rbind(matrix(0, nrow = 4, ncol = 4), c(150, 150*0.05, 150*0.016, 150*0.018))
ii = c(0,10,0.5,0.01)
io = c(0,0.1, 0.1, 0.1)
getPARAMS(intro_comm, externalinputs = ei, inorganicinputs = ii, inorganicloss = io)
}
