% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPARAMS.R
\name{getPARAMS}
\alias{getPARAMS}
\title{A function to get the parameters for a food web model.}
\usage{
getPARAMS(
  usin,
  DIETLIMITS = NA,
  diet_correct = TRUE,
  Conly = FALSE,
  densitydependence = NA,
  functionalresponse = NA,
  externalinputs = NA,
  inorganicinputs = NA,
  inorganicloss = NA,
  forcepositive = FALSE,
  zerosmallchange = NA,
  tracer = NA
)
}
\arguments{
\item{usin}{The community you want to simulate.}

\item{DIETLIMITS}{The diet limits matrix for the stoichiometry correction (proportion of diet)?}

\item{diet_correct}{Boolean: Does the organism correct it's diet?}

\item{Conly}{Boolean: Is the model meant for carbon only?}

\item{densitydependence}{Which nodes have density dependence? NA default means none of them do. Should be a vector of 0 (no DD) and 1 (DD) for each node.}

\item{functionalresponse}{The type of functional response to be used in the model simulation. Either NA to signify a Type I functional response or a matrix of values of the handling time to use a Type II functional response.}

\item{externalinputs}{A matrix of parameters with the rows being the nodes in the food web and the columns being the chemical elements. NA means not inputs except the detritus input necessary to meet food web demand.}

\item{inorganicinputs}{A vector of inputs for the inorganic nutrients. If it is NA, then the inorganic pools are not tracked by the parameter set and the simulation will track net changes in inorganic nutrients by setting inputs and outputs to NA.}

\item{inorganicloss}{A vector of loss rates for the inorganic nutrients. If it is NA, then the inorganic pools are not tracked by the parameter set and the simulation will track net changes in inorganic nutrients by setting inputs and outputs to NA.}

\item{forcepositive}{A TRUE/FALSE parameter that determine if the system of equations zeros any negative pools at each time step. This can be helpful for simulations of complex systems.}

\item{zerosmallchange}{If NA there is no effect. If a value is given, any absolute change less than this is set ot zero. Useful for avoiding drift.}

\item{tracer}{Should a tracer for carbon be included in the model? If default NA, then there is no tracer. Otherwise, a data frame that has the number of rows to match the number of nodes in the food web and two columns. The first column contains the input proportions of carbon, while the second contains the biomass proportions of carbon.}
}
\value{
A list with two elements:(1) a vector of equilibrium biomasses that can be modified and passed to the simulator, and (2) a vector of parameters to run the model away from equilibrium.
}
\description{
A function to get the parameters for a food web model.
}
\details{
A function to get the parameters of a food web model for simulation purposes. It does not correct stoichiometry, so the user must do this beforehand if they want.
For external inputs, Non-zero inputs should match the stoichiometry of non-detritus pools. Detritus inputs should meet demands of the food web if provided.
}
\examples{
# Basic call.
ei = rbind(matrix(0, nrow = 4, ncol = 4), c(150.0,7.5,2.4,2.7), c(15.000,0.450,0.150,0.135))
ii = c(0,10,0.5,0.01)
il = c(0,0.1, 0.1, 0.1)
getPARAMS(intro_comm, externalinputs = ei, inorganicinputs = ii, inorganicloss = il)
# Can also exclude external inputs and have the function calculate the minimum amount necessary.
# You can also exclude inorganic inputs and outputs and
# have the resulting simulation just track changes.
getPARAMS(intro_comm, externalinputs = NA, inorganicinputs = NA, inorganicloss = NA)
}
