% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_function.R
\name{prepare_tracer_inputs}
\alias{prepare_tracer_inputs}
\title{Prepare tracer simulation inputs from main ODE output}
\usage{
prepare_tracer_inputs(
  main_out,
  main_parms,
  tracer_init,
  tracer_input_proportion = NULL
)
}
\arguments{
\item{main_out}{matrix/data.frame from deSolve::ode() using foodwebode WITHOUT tracer}

\item{main_parms}{list of parameters used in the main simulation}

\item{tracer_init}{named numeric vector of initial tracer stocks (length = nrow(main_parms$pmat))}

\item{tracer_input_proportion}{the proportion of tracer in input fluxes}
}
\value{
list with elements:
        - tracer_func: function(t, y, parms) for deSolve
        - tracer_parms: parameter list including main_out and interpolation cache
        - tracer_init: initial tracer state vector
}
\description{
Prepare tracer simulation inputs from main ODE output
}
