% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{removenodes_sim}
\alias{removenodes_sim}
\title{Remove nodes from simulation parameter set.}
\usage{
removenodes_sim(paramset, toremove)
}
\arguments{
\item{paramset}{The parameter set from which to remove nodes.}

\item{toremove}{The name of a node to remove a node using it's name.}
}
\value{
The parameter set without the removed node.
}
\description{
Remove nodes from simulation parameter set.
}
\examples{
removenodes_sim(getPARAMS(intro_comm), c("Pred"))
}
